requirements.txt
pandas
requests
discord.py
aiohttp
